Refactor display_results module
